Support wide lines.
Yet another SPI. I'm still waiting for Apple to make this public after
they briefly promised it would be in Metal 2.3.

Metal's hidden `maxLineWidth` property on `MTLDevice` reports 16 as the
maximum line width, but based on my experiences with the maximum point
size, which actually varies by device from the 511 advertised both in
the Metal Feature Set Tables and by the hidden `maxPointSize` property
(cf. FB8809202), I think it best to leave it at the Vulkan-mandated
minimum of 8.